feat: new endpoints and authentication routes

+++ BalancedLife.Application/Interfaces/IUserService.cs
@@ -3,7 +3,7 @@
 namespace BalancedLife.Application.interfaces {
     public interface IUserService {
         Task<UserInfoDTO> Add(UserDTO user);
-        Task<UserInfoDTO> Update(int id, UserDTO user);
+        Task<UserInfoDTO> Update(long id, UserDTO user);
         Task<UserInfoDTO> GetUserById(int id);
     }
 }
